requirements.txt
google-cloud-bigquery~=3.11.4
openai~=0.27.4
streamlit~=1.24.1
pandas~=1.5.3